requirement.txt
requests
beautifulsoup4
googlesearch-python

tqdm
